Fixes the flow when the conversation has expired
## What
The back end returns a 400 error when the selected conversation has expired. An error toaster is shown, but the user cannot continue chatting until they refresh the page.

## Why
The UI was implemented to support only one chat at a time.

## How
Added a handler for 400 errors so that when this occurs, a new chat is automatically started.