Updates the Docker and deployment resources
- Updated the Dockerfile and other deployment resources in order to add
easier support for providing and configuring a `context path` to the
application.

- The Dockerfile now does additional placeholder replacement in the
`chatbot-ui.conf`, when the file is copied.

- Updated the Helm values to include configuration for the application
`contextPath`. The configuration is used in the configmap which creates
the additional proxy configuration for the backend service.

- Fixed the issue with the query encoding when using the open in SPARQL
editor button, which opens the GraphDB Workbench and places the query
there.

- Fixed the redirect to the GraphDB Workbench. We've added additional
request proxying which strips the `/chat` context path.

- Increased the width of the chat panel so that it can use up the free
space in the page.