Updates couple of deployment resources in order to fix proxying issues
- Updated the default Nginx configuration for the web application. We've
added optional includes of configuration snippet files, which can be
used to define specific proxying/routing to the correct address of the
backend service API.

  This is needed, because currently the web application works with
relative paths in order to invoke the backend service. However, when the
applications are deployed for example in Kubernetes, the backend service
is external for the web application container.

- Updated the `chatbot-service-proxy.conf` to contain only a snippet
that can be added to the default web application Nginx configuration.
Also, now the files is mounted in specific directory for includes.