SN-184 Adds couple of modification to the chatbot service proxy config
- Removed the `$upstream` usage and variable. For some reason it does
not work as expected and described in the documentation (or at least it
doesn't work for our use case).

  Now we are using the address of the chatbot backend service directly.
However, this brings unwanted behavior, when starting the UI container
and the backend service is not accessible.

- Added property for the GraphDB URL in the `configuration` section in
`values.yaml`. The value will be used to create the redirect to the
GraphDB Workbench, when the open in GrahpDB button is clicked for the
SPARQL queries returned by the chatbot.