SN-195: UI to send "X-User-Datetime" header
## What
Added a new header "X-User-Datetime" to all backend calls from the UI.

## Why
The backend needs to know the local time and timezone of the user interacting with the chatbot.

## How
Added an interceptor that automatically adds the "X-User-Datetime" header to all requests.
The header value follows the pattern:
YYYY-MM-DDTHH:mm:ss±HHMM
YYYY – year
- MM – month (01–12)
- DD – day of month (01–31)
- HH – hours (00–23)
- mm – minutes (00–59)
- ss – seconds (00–59)
- ±HH:MM – numeric timezone offset from UTC

Example: 2025-10-09T14:47:19+0300