Adds capability for context path provisioning via environment variable
- We've changed the way `default.conf` for the Nginx is set. Now instead
of having the file directly copied into the `nginx/conf.d/` directory,
the file is provided as `default.conf.template`, which is processed by
initialization scripts, when the Nginx is started.

  This change is need in order to be able to provide the application
context path as environment variable, which is resolved and set in the
template, before creating the actual `default.conf` file into the
`nginx/conf.d/` directory.

  There are some changes in the configurations as well. Instead of
having the `/usr/share/nginx/html/` as root, now we defined it as an
alias.

- We've added custom initialization script that is added to the Nginx.
It will handle the setting of the new context path for the application,
when it is changed. The script is called `context-path-adapter.sh`,
which is added to the Nginx `/docker-entrypoint.d/` directory.

- The Helm chart was updated in order to provide the value of the
`configuration.contextPath` as environment variable to the container.