Adds writable default volumes
- Added volumes configurations, which are providing write permissions
over specific directories in the container. These directories have to be
writable, because Nginx adds specific files in them, on start up.

  Each volume configuration can be disabled. However, you need to make
the directories writable in a different way, either by providing
different volume configurations via `extraVolumes` and
`extraVolumeMounts` or init containers, or proving more relaxed security
context when that is possible.

- Added checksum annotation for the `chatbot-services-proxy-config.yaml`
ConfigMap in order to enable pod updates, when the map content changes.