248 Adds logic for trailing slashes handling when opening the Chatbot UI
- Updated the `default.conf.template` to handle redirects to the
application path, when the user does not enter the trailing slash, which
apparently is required. Otherwise, the Nginx responds with page not
found.

- Also, we've introduced InitContainer to the Helm chart, which allows
us to modify the `index.html` file in environments with elevated
security.

  The issue appeared on the Statnett RNDP environment, where the
container volumes are read-only. However, we have initialization script
for the Nginx container, which does specific modification to some of the
web resources in order to adapt the context path of the application.

- Added additional check to the `context-path-adapter.sh` script, which
makes required the trailing slash, when providing different context
paths to the application via the CONTEXT_PATH environment variable.